# docs/services/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CORTX Gateway
  version: 0.1.0
paths:
...
  /analytics/comparison/summary:
    get:
      tags:
      - Analytics
      summary: Get Comparison Summary
      description: Get high-level comparison statistics between JSON and RAG validation.
      operationId: get_comparison_summary_analytics_comparison_summary_get
      parameters:
      - name: domain
        in: query
        required: false
        schema:
          type: string
          description: Filter by domain
          title: Domain
        description: Filter by domain
      - name: days_back
        in: query
        required: false
        schema:
          type: integer
          description: Number of days to include
          default: 30
          title: Days Back
        description: Number of days to include
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Comparison Summary Analytics Comparison Summary
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analytics/comparison/detailed:
    get:
      tags:
      - Analytics
      summary: Get Detailed Comparison
      description: Get detailed comparison data for specific domain and filters.
      operationId: get_detailed_comparison_analytics_comparison_detailed_get
      parameters:
      - name: domain
        in: query
        required: true
        schema:
          type: string
          description: Domain to analyze
          title: Domain
        description: Domain to analyze
      - name: rule_category
        in: query
        required: false
        schema:
          type: string
          description: Filter by rule category
          title: Rule Category
        description: Filter by rule category
      - name: confidence_min
        in: query
        required: false
        schema:
          type: number
          description: Minimum confidence filter
          default: 0.0
          title: Confidence Min
        description: Minimum confidence filter
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          description: Maximum number of comparisons to return
          default: 100
          title: Limit
        description: Maximum number of comparisons to return
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Detailed Comparison Analytics Comparison Detailed
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analytics/comparison/trends:
    get:
      tags:
      - Analytics
      summary: Get Comparison Trends
      description: Get time-series trends for comparison metrics.
      operationId: get_comparison_trends_analytics_comparison_trends_get
      parameters:
      - name: domain
        in: query
        required: false
        schema:
          type: string
          description: Filter by domain
          title: Domain
        description: Filter by domain
      - name: metric
        in: query
        required: false
        schema:
          type: string
          description: Metric to trend
          default: agreement_rate
          title: Metric
        description: Metric to trend
      - name: period
        in: query
        required: false
        schema:
          type: string
          description: 'Period: hourly, daily, weekly'
          default: daily
          title: Period
        description: 'Period: hourly, daily, weekly'
      - name: days_back
        in: query
        required: false
        schema:
          type: integer
          description: Number of days to include
          default: 30
          title: Days Back
        description: Number of days to include
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Comparison Trends Analytics Comparison Trends
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analytics/comparison/rules/{rule_id}:
    get:
      tags:
      - Analytics
      summary: Get Rule Comparison Analysis
      description: Get detailed analysis for a specific rule's JSON vs RAG performance.
      operationId: get_rule_comparison_analysis_analytics_comparison_rules__rule_id__get
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          title: Rule Id
      - name: days_back
        in: query
        required: false
        schema:
          type: integer
          description: Number of days to analyze
          default: 30
          title: Days Back
        description: Number of days to analyze
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Rule Comparison Analysis Analytics Comparison
                  Rules  Rule Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError